+++ R/modelStudio.R
@@ -57,47 +57,68 @@
 #'                      family = "binomial")
 #'
 #' # Wrap it into an explainer
-#' explain_titanic <- explain(model_titanic,
-#'                            data = titanic_imputed[,-8],
-#'                            y = titanic_imputed[,8],
-#'                            label = "glm",
-#'                            verbose = FALSE)
+#' explainer_titanic <- explain(model_titanic,
+#'                              data = titanic_imputed[,-8],
+#'                              y = titanic_imputed[,8],
+#'                              label = "glm",
+#'                              verbose = FALSE)
 #'
 #' # Pick some data points
 #' new_observations <- titanic_imputed[1:2,]
 #' rownames(new_observations) <- c("Lucas","James")
 #'
 #' # Make a studio for the model
-#' modelStudio(explain_titanic, new_observations,
+#' modelStudio(explainer_titanic, new_observations,
 #'             N = 100, B = 10, show_info = FALSE)
 #'
 #' \donttest{
+#'
 #' #:# ex2 regression on 'apartments' dataset
+#' library("randomForest")
 #'
-#' model_apartments <- randomForest::randomForest(m2.price ~. ,
-#'                                                data = apartments)
+#' model_apartments <- randomForest(m2.price ~. ,data = apartments)
 #'
-#' explain_apartments <- explain(model_apartments,
-#'                               data = apartments[,-1],
-#'                               y = apartments[,1],
-#'                               verbose = FALSE)
+#' explainer_apartments <- explain(model_apartments,
+#'                                 data = apartments[,-1],
+#'                                 y = apartments[,1],
+#'                                 verbose = FALSE)
 #'
 #' new_apartments <- apartments[1:2,]
 #' rownames(new_apartments) <- c("ap1","ap2")
 #'
 #' # change dashboard dimensions and animation length
-#' modelStudio(explain_apartments, new_apartments,
+#' modelStudio(explainer_apartments, new_apartments,
 #'             facet_dim = c(2, 3), time = 800,
 #'             show_info = FALSE)
 #'
 #' # add information about true labels
-#' modelStudio(explain_apartments, new_apartments,
+#' modelStudio(explainer_apartments, new_apartments,
 #'                                 new_observation_y = apartments[1:2, 1],
 #'                                 show_info = FALSE)
 #'
 #' # don't compute EDA plots
-#' modelStudio(explain_apartments, eda = FALSE,
+#' modelStudio(explainer_apartments, eda = FALSE,
 #'             show_info = FALSE)
+#'
+#'
+#' #:# ex3 xgboost model on 'HR' dataset
+#' library("xgboost")
+#'
+#' model_matrix <- model.matrix(status == "fired" ~ . -1, HR)
+#' data <- xgb.DMatrix(model_matrix, label = HR$status == "fired")
+#'
+#' params <- list(max_depth = 2, eta = 1, silent = 1, nthread = 2,
+#'                objective = "binary:logistic", eval_metric = "auc")
+#'
+#' model_HR <- xgb.train(params, data, nrounds = 50)
+#'
+#' explainer_HR <- explain(model_HR,
+#'                         data = model_matrix,
+#'                         y = HR$status == "fired",
+#'                         verbose = FALSE)
+#'
+#' modelStudio(explainer_HR, show_info = FALSE)
+#'
 #' }
 #'
 #' @export
@@ -129,18 +150,18 @@ modelStudio.explainer <- function(explainer,
   predict_function <- explainer$predict_function
   label <- explainer$label
 
-  if (is.null(new_observation) & show_info) {
-    message("`new_observation` argument is NULL. Observations needed to calculate local explanations are taken at random from the data.")
+  if (is.null(new_observation)) {
+    if (show_info) message("`new_observation` argument is NULL.\nObservations needed to calculate local explanations are taken at random from the data.")
     new_observation <- ingredients::select_sample(data, 3)
   }
 
   ## safeguard
-  #new_observation <- as.data.frame(new_observation)
-  #data <- as.data.frame(data)
-
   if (is.null(dim(new_observation))) {
     warning("`new_observation` argument is not a data.frame nor a matrix, coerced to data.frame")
     new_observation <- as.data.frame(new_observation)
+  }
+  if (is.null(rownames(data))) {
+    rownames(data) <- 1:nrow(data)
   }
 
   ## get proper names of features that arent target
@@ -148,7 +169,7 @@ modelStudio.explainer <- function(explainer,
   potential_variable_names <- names(is_y[!is_y])
   variable_names <- intersect(potential_variable_names, colnames(new_observation))
   ## get rid of target in data
-  data <- data[!is_y]
+  data <- data[,!is_y]
 
   obs_count <- dim(new_observation)[1]
   obs_data <- new_observation
@@ -165,7 +186,7 @@ modelStudio.explainer <- function(explainer,
 
   if (show_info) setTxtProgressBar(pb, 1)
 
-  which_numerical <- sapply(data[,, drop = FALSE], is.numeric)
+  which_numerical <- apply(data[,, drop = FALSE], 2, is.numeric)
 
   ## because aggregate_profiles calculates numerical OR categorical
   if (all(which_numerical)) {
@@ -233,7 +254,7 @@ modelStudio.explainer <- function(explainer,
     parallelMap::parallelLibrary(packages = loadedNamespaces())
 
     f <- function(i, model, data, predict_function, label, B, show_boxplot, ...) {
-      new_observation <- obs_data[i,]
+      new_observation <- obs_data[i, , drop = FALSE]
 
       bd <- try_catch(
         iBreakDown::local_attributions(
@@ -272,7 +293,7 @@ modelStudio.explainer <- function(explainer,
   } else {
     ## count once per observation
     for(i in 1:obs_count){
-      new_observation <- obs_data[i,]
+      new_observation <- obs_data[i, , drop = FALSE]
 
       bd <- try_catch(
         iBreakDown::local_attributions(